+++ server/src/main/java/com/linbit/linstor/event/common/DonePercentageEvent.java
@@ -2,6 +2,7 @@ package com.linbit.linstor.event.common;
 
 import com.linbit.linstor.event.GenericEvent;
 import com.linbit.linstor.event.LinstorTriggerableEvent;
+import com.linbit.utils.Pair;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -11,15 +12,15 @@ import java.util.Optional;
 @Singleton
 public class DonePercentageEvent
 {
-    private final LinstorTriggerableEvent<Optional<Float>> event;
+    private final LinstorTriggerableEvent<Pair<String, Optional<Float>>> event;
 
     @Inject
-    public DonePercentageEvent(GenericEvent<Optional<Float>> eventRef)
+    public DonePercentageEvent(GenericEvent<Pair<String, Optional<Float>>> eventRef)
     {
         event = eventRef;
     }
 
-    public LinstorTriggerableEvent<Optional<Float>> get()
+    public LinstorTriggerableEvent<Pair<String, Optional<Float>>> get()
     {
         return event;
     }