+++ server/src/main/java/com/linbit/linstor/event/common/ReplicationStateEvent.java
@@ -3,6 +3,7 @@ package com.linbit.linstor.event.common;
 import com.linbit.linstor.event.GenericEvent;
 import com.linbit.linstor.event.LinstorTriggerableEvent;
 import com.linbit.linstor.layer.drbd.drbdstate.ReplState;
+import com.linbit.utils.Pair;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -10,15 +11,15 @@ import javax.inject.Singleton;
 @Singleton
 public class ReplicationStateEvent
 {
-    private final LinstorTriggerableEvent<ReplState> event;
+    private final LinstorTriggerableEvent<Pair<String, ReplState>> event;
 
     @Inject
-    public ReplicationStateEvent(GenericEvent<ReplState> eventRef)
+    public ReplicationStateEvent(GenericEvent<Pair<String, ReplState>> eventRef)
     {
         event = eventRef;
     }
 
-    public LinstorTriggerableEvent<ReplState> get()
+    public LinstorTriggerableEvent<Pair<String, ReplState>> get()
     {
         return event;
     }