+++ server/src/main/java/com/linbit/linstor/satellitestate/SatelliteVolumeState.java
@@ -1,12 +1,25 @@
 package com.linbit.linstor.satellitestate;
 
 import com.linbit.linstor.layer.drbd.drbdstate.ReplState;
+import com.linbit.utils.Pair;
+
+import javax.annotation.Nullable;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
 
 public class SatelliteVolumeState
 {
     private String diskState;
-    private ReplState replicationState;
-    private Float donePercentage;
+    /**
+     * Stores the replication states for each peer connection, key is the peer name, NOT the node name
+     */
+    private Map<String, ReplState> replicationStateMap = new HashMap<>();
+    /**
+     * Stores the done percentage for each peer connection, key is the peer name, NOT the node name
+     */
+    private Map<String, Float> donePercentageMap = new HashMap<>();
 
     public SatelliteVolumeState()
     {
@@ -15,8 +28,8 @@ public class SatelliteVolumeState
     public SatelliteVolumeState(SatelliteVolumeState other)
     {
         diskState = other.diskState;
-        replicationState = other.replicationState;
-        donePercentage = other.donePercentage;
+        replicationStateMap = other.replicationStateMap;
+        donePercentageMap = other.donePercentageMap;
     }
 
     public String getDiskState()
@@ -29,28 +42,56 @@ public class SatelliteVolumeState
         diskState = diskStateRef;
     }
 
-    public ReplState getReplicationState()
+    public Map<String, ReplState> getReplicationStateMap()
     {
-        return replicationState;
+        return replicationStateMap;
     }
 
-    public void setReplicationState(ReplState replicationStateRef)
+    public void setReplicationState(@Nullable Pair<String, ReplState> replStatePair)
     {
-        this.replicationState = replicationStateRef;
+        if (replStatePair != null)
+        {
+            if (replStatePair.objB == null || replStatePair.objB == ReplState.OFF)
+            {
+                replicationStateMap.remove(replStatePair.objA);
+            }
+            else
+            {
+                replicationStateMap.put(replStatePair.objA, replStatePair.objB);
+            }
+        }
+        else
+        {
+            replicationStateMap.clear();
+        }
     }
 
-    public Float getDonePercentage()
+    public Map<String, Float> getDonePercentageMap()
     {
-        return donePercentage;
+        return donePercentageMap;
     }
 
-    public void setDonePercentage(Float donePercentageRef)
+    public void setDonePercentage(@Nullable Pair<String, Optional<Float>> donePercentagePair)
     {
-        donePercentage = donePercentageRef;
+        if (donePercentagePair != null)
+        {
+            if (donePercentagePair.objB.isEmpty())
+            {
+                donePercentageMap.remove(donePercentagePair.objA);
+            }
+            else
+            {
+                donePercentageMap.put(donePercentagePair.objA, donePercentagePair.objB.get());
+            }
+        }
+        else
+        {
+            donePercentageMap.clear();
+        }
     }
 
     public boolean isEmpty()
     {
-        return diskState == null && replicationState == null && donePercentage == null;
+        return diskState == null && replicationStateMap.isEmpty() && donePercentageMap.isEmpty();
     }
 }